refactor(planner): Move data interfacing logic to custom react hooks
In "app/plans/[...planId].tsx, the process of obtaining the user's plan and
loading in the plan state has been moved to the usePlan hook. Additionally, the
logic for calling courseAttempts has been moved to its own file.
